Split and move sort_inputs.rs into sort_inputs/ module
