+++ src/uu/sort/src/sort.rs
@@ -17,7 +17,7 @@ mod custom_str_cmp;
 mod ext_sort;
 mod merge;
 mod numeric_str_cmp;
-mod sort_input;
+mod sort_inputs;
 mod tmp_dir;
 
 use bigdecimal::BigDecimal;
