+++ src/uu/sort/src/sort_inputs/mod.rs
@@ -1,0 +1,4 @@
+mod iterator;
+mod model;
+
+pub use model::{SortInput, SortInputInner, SortInputs};
